Implements wings suggestions animation
